requirements.txt
django
antlr4-python3-runtime
antlr4-tools
flask
flask_cors